feat: Add support for sorting and filtering blog posts by user
The code changes in the `ControllerBlogPost` class and `ServiceBlogPost` class introduce support for sorting and filtering blog posts based on the user. The `listAll` method now accepts an additional parameter `userId` to filter the posts by user. The `sorted` parameter is used to determine the sorting order of the posts. This update improves the functionality of the blog post listing feature.