feat(front-end): Update BlogPost component to fix user ownership check
The BlogPost component has been updated to fix the user ownership check. Previously, the check was comparing the user ID of the logged-in user with the user ID of the blog post, which caused an error when the blog post did not have a user object. The code has been modified to compare the user ID with the userId property of the blog post instead. This ensures that the ownership check works correctly in all scenarios.